Refactor: core navigation
