+++ app/core/src/navigation/index.js
@@ -1,57 +1,24 @@
 // @flow
 
-import * as React from 'react';
-import {
-  StackNavigator,
-  StackNavigatorOptions,
-} from '@kiwicom/mobile-navigation';
 import { Color } from '@kiwicom/mobile-shared';
-import { Translation } from '@kiwicom/mobile-localization';
-import ProfileScreen from '@kiwicom/mobile-profile';
 import { createBottomTabNavigator } from 'react-navigation';
 
-import HotelsStack from './HotelsStack';
-import MMBPackage from '../screens/MMBPackageWrapper';
-
-const VoidStack = StackNavigator(
-  {
-    Void: {
-      screen: function VoidScreen() {
-        return (
-          <Translation passThrough="This scene is prepared for the future. Please check navigation in the core package." />
-        );
-      },
-    },
-  },
-  {
-    ...StackNavigatorOptions,
-    initialRouteName: 'Void',
-  },
-);
-
-const ProfileStack = StackNavigator(
-  {
-    Profile: {
-      screen: ProfileScreen,
-    },
-  },
-  {
-    ...StackNavigatorOptions,
-    initialRouteName: 'Profile',
-  },
-);
+import HotelsStack from './stacks/HotelsStack';
+import VoidStack from './stacks/VoidStack';
+import ProfileStack from './stacks/ProfileStack';
+import BookingsStack from './stacks/BookingsStack';
 
 export default createBottomTabNavigator(
   {
     Search: { screen: HotelsStack },
-    Bookings: { screen: MMBPackage },
+    Bookings: { screen: BookingsStack },
     Message: { screen: VoidStack },
     Profile: { screen: ProfileStack },
   },
   {
     tabBarOptions: {
       activeTintColor: Color.brand,
-      inactiveTintColor: 'gray',
+      inactiveTintColor: Color.grey.$600,
       style: {
         backgroundColor: Color.white, // makes Android tab bar white instead of standard blue
       },
